requirements.txt
begins==0.9
certifi==2018.1.18
chardet==3.0.4
idna==2.6
PyJWT==1.6.1
PySocks==1.6.8
pytz==2018.4
requests==2.18.4
six==1.11.0
twilio==6.12.0
urllib3==1.22
